+++ src/Storage/Storage.Management.Test/ScenarioTests/StorageBlobTests.ps1
@@ -33,7 +33,7 @@ function Test-StorageBlobIsVersioningEnabled
         $kind = 'StorageV2'
 	
         Write-Verbose "RGName: $rgname | Loc: $loc"
-        New-AzResourceGroup -Name $rgname -Location $loc;
+        New-AzResourceGroup -Name $rgname -Location $loc -Tag @{Some = 'some'};
 		
         $loc = Get-ProviderLocation_Canary ResourceManagement;
         New-AzStorageAccount -ResourceGroupName $rgname -Name $stoname -Location $loc -Type $stotype -Kind $kind